requirements.txt
streamlit
pillow
pickle5
torch
torchvision